fast_api/requirements.txt
fastapi
mysql-connector-python
pydantic
python-dotenv
uvicorn
PyJWT
slowapi
limits